+++ h3ron-polars/src/spatial_index/kdtree.rs
@@ -1,3 +1,6 @@
+use crate::spatial_index::{
+    finish_mask, negative_mask, CoordinateIndexable, CoordinateSIKind, SpatialIndex,
+};
 use crate::{AsH3IndexChunked, Error, IndexChunked, IndexValue};
 use geo::{BoundingRect, Contains};
 use geo_types::{Coordinate, MultiPolygon, Polygon, Rect};
@@ -10,28 +13,9 @@ use polars::prelude::{BooleanChunked, TakeRandom};
 use polars_core::prelude::UInt64Chunked;
 use std::marker::PhantomData;
 
-pub trait KDBushCoordinate {
-    fn kdbush_coordinate(&self) -> Result<Coordinate, Error>;
-}
-
-impl KDBushCoordinate for H3Cell {
-    fn kdbush_coordinate(&self) -> Result<Coordinate, Error> {
-        Ok(self.to_coordinate()?)
-    }
-}
-
-impl KDBushCoordinate for H3DirectedEdge {
-    fn kdbush_coordinate(&self) -> Result<Coordinate, Error> {
-        let cells = self.cells()?;
-        let c1 = cells.destination.to_coordinate()?;
-        let c2 = cells.origin.to_coordinate()?;
-        Ok(((c1.x + c2.x) / 2.0, (c1.y + c2.y) / 2.0).into())
-    }
-}
-
 impl<'a, IX> PointReader for IndexChunked<'a, IX>
 where
-    IX: IndexValue + KDBushCoordinate,
+    IX: IndexValue + CoordinateIndexable,
 {
     fn size_hint(&self) -> usize {
         self.len()
@@ -43,7 +27,7 @@ where
     {
         for (id, maybe_point) in self.iter_indexes_validated().enumerate() {
             if let Some(Ok(point)) = maybe_point {
-                if let Ok(coord) = point.kdbush_coordinate() {
+                if let Ok(coord) = point.spatial_index_coordinate() {
                     visitor(id, coord.x, coord.y);
                 }
             }
@@ -51,11 +35,11 @@ where
     }
 }
 
-pub trait BuildKDBushIndex<'a, IX>
+pub trait BuildKDTreeIndex<'a, IX>
 where
-    IX: IndexValue + KDBushCoordinate,
+    IX: IndexValue + CoordinateIndexable,
 {
-    /// Build a [`KDBushIndex`] using the default parameters.
+    /// Build a [`KDTreeIndex`] using the default parameters.
     ///
     /// # Example
     ///
@@ -65,7 +49,7 @@ where
     /// use polars_core::prelude::TakeRandom;
     /// use h3ron::{H3Cell, Index};
     /// use h3ron_polars::{AsH3CellChunked, NamedFromIndexes};
-    /// use h3ron_polars::spatial_index::BuildKDBushIndex;
+    /// use h3ron_polars::spatial_index::{BuildKDTreeIndex, SpatialIndex, SpatialIndexGeomOp};
     ///
     /// let ca = UInt64Chunked::new_from_indexes(
     ///     "",
@@ -76,31 +60,31 @@ where
     ///     ],
     /// );
     ///
-    /// let kdbush = ca.h3cell().kdbush_index();
-    /// let mask = kdbush.within_rect(&Rect::new((40.0, 40.0), (50.0, 50.0)));
+    /// let kdbush = ca.h3cell().kdtree_index();
+    /// let mask = kdbush.geometries_intersect(&Rect::new((40.0, 40.0), (50.0, 50.0)));
     ///
     /// assert_eq!(mask.len(), 3);
     /// assert_eq!(mask.get(0), Some(true));
     /// assert_eq!(mask.get(1), Some(false));
     /// assert_eq!(mask.get(2), Some(false));
     /// ```
-    fn kdbush_index(&self) -> KDBushIndex<IX> {
-        self.kdbush_index_with_node_size(64)
+    fn kdtree_index(&self) -> KDTreeIndex<IX> {
+        self.kdtree_index_with_node_size(64)
     }
 
-    /// Build a [`KDBushIndex`] using custom parameters.
+    /// Build a [`KDTreeIndex`] using custom parameters.
     ///
     /// `node_size` - Size of the KD-tree node, 64 by default. Higher means faster indexing but slower search, and vise versa
-    fn kdbush_index_with_node_size(&self, node_size: u8) -> KDBushIndex<IX>;
+    fn kdtree_index_with_node_size(&self, node_size: u8) -> KDTreeIndex<IX>;
 }
 
-impl<'a, IX: IndexValue> BuildKDBushIndex<'a, IX> for IndexChunked<'a, IX>
+impl<'a, IX: IndexValue> BuildKDTreeIndex<'a, IX> for IndexChunked<'a, IX>
 where
     Self: PointReader,
-    IX: KDBushCoordinate,
+    IX: CoordinateIndexable,
 {
-    fn kdbush_index_with_node_size(&self, node_size: u8) -> KDBushIndex<IX> {
-        KDBushIndex {
+    fn kdtree_index_with_node_size(&self, node_size: u8) -> KDTreeIndex<IX> {
+        KDTreeIndex {
             index_phantom: PhantomData::<IX>::default(),
 
             // clones of arrow-backed arrays are cheap, so we clone this for the benefit of not
@@ -115,33 +99,24 @@ where
 /// A very fast static spatial index for 2D points based on a flat KD-tree.
 ///
 /// Operates on the centroid coordinate of [`H3Cell`] and [`H3DirectedEdge`] values.
-pub struct KDBushIndex<IX: IndexValue> {
+pub struct KDTreeIndex<IX: IndexValue> {
     index_phantom: PhantomData<IX>,
 
     chunked_array: UInt64Chunked,
 
-    kdbush: KDBush,
+    pub kdbush: KDBush,
 }
 
-impl<IX: IndexValue> KDBushIndex<IX>
+impl<IX: IndexValue> SpatialIndex<IX, CoordinateSIKind> for KDTreeIndex<IX>
 where
-    IX: KDBushCoordinate,
+    IX: CoordinateIndexable,
 {
-    fn build_negative_mask(&self) -> MutableBitmap {
-        let mut mask = MutableBitmap::new();
-        mask.extend_constant(self.chunked_array.len(), false);
-        mask
+    fn h3indexchunked(&self) -> IndexChunked<IX> {
+        self.chunked_array.h3indexchunked()
     }
 
-    fn finish_mask(&self, mask: Bitmap) -> BooleanChunked {
-        let index_chunked = self.chunked_array.h3indexchunked::<IX>();
-        let validites = index_chunked.validity_bitmap();
-        let bool_arr = BooleanArray::from_data(ArrowDataType::Boolean, mask, Some(validites));
-        BooleanChunked::from(bool_arr)
-    }
-
-    fn within_rect_impl(&self, rect: &Rect) -> MutableBitmap {
-        let mut mask = self.build_negative_mask();
+    fn envelopes_intersect_impl(&self, rect: &Rect) -> MutableBitmap {
+        let mut mask = negative_mask(&self.chunked_array);
         self.kdbush.range(
             rect.min().x,
             rect.min().y,
@@ -152,71 +127,19 @@ where
         mask
     }
 
-    pub fn within_rect(&self, rect: &Rect) -> BooleanChunked {
-        self.finish_mask(self.within_rect_impl(rect).into())
-    }
-
-    pub fn within_distance(&self, coord: Coordinate, distance: f64) -> BooleanChunked {
-        let mut mask = self.build_negative_mask();
+    fn envelopes_within_distance(&self, coord: Coordinate, distance: f64) -> BooleanChunked {
+        let mut mask = negative_mask(&self.chunked_array);
         self.kdbush
             .within(coord.x, coord.y, distance, |id| mask.set(id, true));
-        self.finish_mask(mask.into())
-    }
-
-    pub fn within_polygon(&self, polygon: &Polygon) -> Result<BooleanChunked, Error> {
-        let mask =
-            index_mask_within_polygon(self, polygon).unwrap_or_else(|| self.build_negative_mask());
-        Ok(self.finish_mask(mask.into()))
-    }
-
-    pub fn within_multipolygon(
-        &self,
-        multipolygon: &MultiPolygon,
-    ) -> Result<BooleanChunked, Error> {
-        let mask = multipolygon
-            .0
-            .iter()
-            .map(|poly| index_mask_within_polygon(self, poly))
-            .reduce(|a, b| match (a, b) {
-                (None, Some(b)) => Some(b),
-                (Some(a), None) => Some(a),
-                (None, None) => None,
-                (Some(a), Some(b)) => Some(a | &(b.into())),
-            })
-            .flatten()
-            .unwrap_or_else(|| self.build_negative_mask());
-        Ok(self.finish_mask(mask.into()))
-    }
-}
-
-fn index_mask_within_polygon<IX>(kdbi: &KDBushIndex<IX>, polygon: &Polygon) -> Option<MutableBitmap>
-where
-    IX: KDBushCoordinate + IndexValue,
-{
-    if let Some(rect) = polygon.bounding_rect() {
-        let mut mask = kdbi.within_rect_impl(&rect);
-
-        let indexchunked = kdbi.chunked_array.h3indexchunked::<IX>();
-        for i in 0..mask.len() {
-            if mask.get(i) {
-                if let Some(coord) = indexchunked
-                    .get(i)
-                    .and_then(|cell| cell.kdbush_coordinate().ok())
-                {
-                    mask.set(i, polygon.contains(&coord))
-                }
-            }
-        }
-        Some(mask)
-    } else {
-        None
+        finish_mask(mask.into(), &self.h3indexchunked())
     }
 }
 
 #[cfg(test)]
 mod test {
     use crate::from::NamedFromIndexes;
-    use crate::spatial_index::kdbush::BuildKDBushIndex;
+    use crate::spatial_index::kdtree::BuildKDTreeIndex;
+    use crate::spatial_index::{SpatialIndex, SpatialIndexGeomOp};
     use crate::AsH3CellChunked;
     use geo_types::{coord, polygon, Rect};
     use h3ron::{H3Cell, Index};
@@ -237,8 +160,8 @@ mod test {
     #[test]
     fn cell_within_distance() {
         let ca = build_cell_ca();
-        let kd = ca.h3cell().kdbush_index();
-        let mask = kd.within_distance((-60.0, -60.0).into(), 2.0);
+        let kd = ca.h3cell().kdtree_index();
+        let mask = kd.envelopes_within_distance((-60.0, -60.0).into(), 2.0);
 
         assert_eq!(mask.len(), 4);
         assert_eq!(mask.get(0), Some(false));
@@ -250,8 +173,8 @@ mod test {
     #[test]
     fn cell_within_rect() {
         let ca = build_cell_ca();
-        let kd = ca.h3cell().kdbush_index();
-        let mask = kd.within_rect(&Rect::new((40.0, 40.0), (50.0, 50.0)));
+        let kd = ca.h3cell().kdtree_index();
+        let mask = kd.geometries_intersect(&Rect::new((40.0, 40.0), (50.0, 50.0)));
 
         assert_eq!(mask.len(), 4);
         assert_eq!(mask.get(0), Some(true));
@@ -263,16 +186,14 @@ mod test {
     #[test]
     fn cell_within_polygon() {
         let ca = build_cell_ca();
-        let kd = ca.h3cell().kdbush_index();
-        let mask = kd
-            .within_polygon(&polygon!(exterior: [
+        let kd = ca.h3cell().kdtree_index();
+        let mask = kd.geometries_intersect_polygon(&polygon!(exterior: [
                 coord! {x: 40.0, y: 40.0},
                 coord! {x: 40.0, y: 50.0},
                 coord! {x: 49.0, y: 50.0},
                 coord! {x: 49.0, y: 40.0},
                 coord! {x: 40.0, y: 40.0},
-            ], interiors: []))
-            .unwrap();
+            ], interiors: []));
 
         assert_eq!(mask.len(), 4);
         assert_eq!(mask.get(0), Some(true));