+++ h3ron/CHANGES.md
@@ -14,6 +14,7 @@ python extension.
 ### Added
 * Add example describing how to compile to the wasm32-wasi target.
 * Parse `H3Cell` from different string representations using `H3Cell::from_str`. [#52](https://github.com/nmandery/h3ron/pull/52)
+* `ToLine` trait and implementation for `H3DirectedEdge`.
 
 ### Changed
 * Use only types from `geo-types` in the public API, not from `geo`.